Rename tor_connection_dialog.py to tor_connection.py
